requirements.txt
langchain>=0.1.0
openai>=1.3.0
python-dotenv>=1.0.0
streamlit>=1.30.0
fastapi>=0.110.0
uvicorn>=0.25.0
selenium>=4.10.0
chromadb>=0.4.22
faiss-cpu>=1.7.4
